+++ Backend-Supermercado/Backend-Supermercado-ejb/src/java/org/back/ejb/GestionSupermercadoEjbLocal.java
@@ -19,7 +19,7 @@ public interface GestionSupermercadoEjbLocal {
     
     List<Supermercado> listarSupermercados() throws Exception;
 
-    Supermercado buscarSupermercado(long idSupermercado) throws Exception;
+    Supermercado buscarSupermercado(Integer idSupermercado) throws Exception;
 
     Supermercado guardarSupermercado(Supermercado supermercado) throws Exception;
